requirements.txt
# ==========================================
# memory-mcp v2 dependencies
# ==========================================

# Embedding & Reranker
sentence-transformers>=3.0.0
sentencepiece>=0.1.99  # Required for cl-nagoya/ruri-v3-30m tokenizer

# PyTorch CPU版（sentence-transformersの依存として自動インストールされる）
# Docker: Dockerfile で --index-url を使って CPU 版を先にインストール
# ローカル開発: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu

# MCP Server
mcp>=1.0.0

# Data Validation & Settings
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0

# Vector Store
qdrant-client>=1.8.2

# Numerical
numpy>=1.24.0

# E2E Testing (Playwright)
playwright>=1.40.0
pytest-playwright>=0.4.0
